net: wireless: update bcmdhd driver to 101.10.361.22

+++ drivers/net/wireless/rockchip_wlan/rkwifi/bcmdhd/wl_cfgscan.h
@@ -173,6 +173,6 @@ typedef enum {
 
 extern bool wl_cfgscan_is_dfs_set(wifi_band band);
 extern s32 wl_cfgscan_get_band_freq_list(struct bcm_cfg80211 *cfg, int band,
-        uint16 *list, uint32 *num_channels);
+        uint32 *list, uint32 *num_channels);
 #endif /* DHD_GET_VALID_CHANNELS */
 #endif /* _wl_cfgscan_h_ */
